change default-files-when-install.php

+++ default-files-when-install.php
@@ -1,20 +1,25 @@
 <?php
-
 final class Setup {
-    public static function createDirectories(array $directories) : ? string{
+    public static function createDirectories(array $directories) {
         foreach ($directories as $dir) {
             if (!is_dir($dir)) {
-                mkdir($dir, 0755, true);
-                echo "[*] Created directory: $dir\n";
+                if (mkdir($dir, 0755, true)) {
+                    echo "[*] Created directory: $dir\n";
+                } else {
+                    echo "[!] Failed to create directory: $dir\n";
+                }
             }
         }
     }
 
-    public static function createFiles(array $files) : ? string{
+    public static function createFiles(array $files) {
         foreach ($files as $file => $content) {
             if (!file_exists($file)) {
-                file_put_contents($file, $content);
-                echo "[*] Created file: $file\n";
+                if (file_put_contents($file, $content) !== false) {
+                    echo "[*] Created file: $file\n";
+                } else {
+                    echo "[!] Failed to create file: $file\n";
+                }
             }
         }
     }
@@ -27,11 +32,13 @@ $directories = [
 ];
 
 $files = [
-    'config/cool-view.php' =>
-        "<?php\n\nreturn [\n    // Defines the prefix used for template directives (#if, #foreach)\n    'PrefixCharCoolEngine' => \"#\",\n\n    'ViewPath' => \"/template/views\",\n\n    'CachePath' => \"/template/caches\",\n];",
+    'config/cool-view.php' => "<?php\n\nreturn [\n    'PrefixCharCoolEngine' => \"#\",\n    'ViewPath' => \"/template/views\",\n    'CachePath' => \"/template/caches\",\n];",
 ];
 
 Setup::createDirectories($directories);
 Setup::createFiles($files);
 
-unlink(__FILE__); // delete after setup;
+if (file_exists(__FILE__)) {
+    unlink(__FILE__);
+    echo "[*] Setup script deleted.\n";
+}
